+++ src/Helpers/GatewayHelper.php
@@ -29,6 +29,7 @@ use Buckaroo\Shopware6\PaymentMethods\KlarnaKp;
 use Buckaroo\Shopware6\PaymentMethods\Klarnain;
 use Buckaroo\Shopware6\PaymentMethods\Billink;
 use Buckaroo\Shopware6\PaymentMethods\Belfius;
+use Buckaroo\Shopware6\PaymentMethods\PayPerEmail;
 
 class GatewayHelper
 {
@@ -59,6 +60,7 @@ class GatewayHelper
         KlarnaKp::class,
         Klarnain::class,
         Billink::class,
-        Belfius::class
+        Belfius::class,
+        PayPerEmail::class
     ];
 }