+++ src/PaymentMethods/PayPerEmail.php
@@ -1,0 +1,76 @@
+<?php declare(strict_types=1);
+
+namespace Buckaroo\Shopware6\PaymentMethods;
+
+use Buckaroo\Shopware6\Handlers\PayPerEmailPaymentHandler;
+
+class PayPerEmail extends AbstractPayment
+{
+    /*
+    * @return string
+    */
+    public function getBuckarooKey(): string
+    {
+        return 'payperemail';
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getName(): string
+    {
+        return 'Buckaroo PayPerEmail';
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getDescription(): string
+    {
+        return 'Pay with PayPerEmail';
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getMedia(): string
+    {
+        return __DIR__  . '/../Resources/views/storefront/buckaroo/logo/payperemail.png';
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getPaymentHandler(): string
+    {
+        return PayPerEmailPaymentHandler::class;
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return array
+     */
+    public function getTranslations(): array
+    {
+        return [
+            'de-DE' => [
+                'name'        => $this->getName(),
+                'description' => 'Bezahlen mit PayPerEmail',
+            ],
+            'en-GB' => [
+                'name'        => $this->getName(),
+                'description' => $this->getDescription(),
+            ],
+        ];
+    }
+
+}