+++ src/Resources/app/administration/src/api/buckaroo-payment.service.js
@@ -55,6 +55,23 @@ class BuckarooPaymentService extends ApiService {
             return ApiService.handleResponse(response);
         });
     }
+
+    createPaylink(transaction, transactionsToRefund, orderItems) {
+        const apiRoute = `_action/${this.getApiBasePath()}/paylink`;
+
+        return this.httpClient.post(
+            apiRoute,
+            {
+                transaction: transaction
+            },
+            {
+                headers: this.getBasicHeaders()
+            }
+        ).then((response) => {
+            return ApiService.handleResponse(response);
+        });
+    }
+
 }
 
 Application.addServiceProvider('BuckarooPaymentService', (container) => {