+++ src/Resources/app/administration/src/module/buckaroo-payment/page/buckaroo-settings/index.js
@@ -59,8 +59,10 @@ Component.register('buckaroo-settings', {
                 'klarnakpEnabled': true,
                 'advancedConfiguration': true,
                 'BillinkCreateInvoiceAfterShipment': true,
+                'payperemailEnabled': true,
             },
             collapsibleAdvancedState: {
+                'paylinkEnabled': true,
                 'stockReserve': true,
                 'sendInvoiceEmail': true,
                 'showPaymentMethodsList': true,